Added dependencies for play framework and database schema
