+++ conf/application.conf
@@ -1,0 +1,13 @@
+play.http.secret.key = "1JqxAyWlCEGn4f2mZxMtjoJ4uQ0FHuHC"
+server-url = "http://213.239.193.208:9053/"
+best-block-id = "c89d70cf4d7664676dd53e091617d40aa183ba6e0485cd077b7b74e983c77fd2"
+
+slick.dbs.default.profile = "slick.jdbc.H2Profile$"
+slick.dbs.default.driver = "slick.driver.H2Driver$"
+slick.dbs.default.db.driver = "org.h2.Driver"
+slick.dbs.default.db.url = "jdbc:h2:./database/scanner"
+slick.dbs.default.db.user = test
+slick.dbs.default.db.password = "test"
+play.evolutions.autoApply = true
+//slick.dbs.default.db.numThreads=20
+//slick.dbs.default.db.maxConnections=20
