+++ conf/routes
@@ -1,0 +1,4 @@
+GET         /                               controllers.Controller.home
+
+GET         /assets/*file                   controllers.Assets.versioned(path="/public", file: Asset)
+
