+++ project/plugins.sbt
@@ -1,0 +1,4 @@
+logLevel := Level.Warn
+
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.9")
+addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.8.1")
